feat(doctor): warn when competing memory surfaces exist
Extract doctor reporting into internal/doctor and warn when MEMORY.md or Claude native memory coexists with an active .mnemo marker.
