+++ cmd/mnemo/doctor_types.go
@@ -1,0 +1,9 @@
+package main
+
+type doctorOptions struct {
+	JSON    bool
+	Agent   string
+	Path    string
+	Home    string
+	DataDir string
+}
